backend/requirements.txt
aniso8601==10.0.1
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
Flask==3.1.2
Flask-JWT-Extended==4.7.1
flask-restx==1.3.2
Flask-SQLAlchemy==3.1.1
google-auth==2.47.0
greenlet==3.3.0
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jwt==1.4.0
MarkupSafe==3.0.3
numpy==2.4.1
pandas==2.3.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
semantic-version==2.10.0
setuptools==80.9.0
setuptools-rust==1.12.0
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
